Move ObjectSourceGame enum to its own header
